Move tool batch execution into runner
